# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Procurement Standards API",
    "description": "Offline metadata recommendations. Every result carries evidence; legal applicability requires review.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/recommend": {
      "post": {
        "summary": "Recommend",
        "operationId": "recommend_v1_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "top_k": {
                    "type": "integer",
                    "default": 5
                  },
                  "language_hint": {
                    "type": "string"
                  },
                  "product_category": {
                    "type": "string"
                  },
                  "certification_context": {
                    "type": "string",
                    "description": "JSON object of scope assertions"
                  }
                },
                "type": "object",
                "required": [
                  "file"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/system": {
      "get": {
        "summary": "System",
        "operationId": "system_v1_system_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/history": {
      "get": {
        "summary": "History",
        "operationId": "history_v1_history_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/history/{report_id}": {
      "get": {
        "summary": "Saved Report",
        "operationId": "saved_report_v1_history__report_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/standards": {
      "get": {
        "summary": "Directory",
        "operationId": "directory_v1_standards_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/standards/{is_number}/allied": {
      "get": {
        "summary": "Allied",
        "operationId": "allied_v1_standards__is_number__allied_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "is_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Is Number"
            }
          },
          {
            "name": "max_hops",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 4,
              "minimum": 0,
              "default": 2,
              "title": "Max Hops"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlliedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/standards/{is_number}": {
      "get": {
        "summary": "Standard",
        "operationId": "standard_v1_standards__is_number__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "is_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Is Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/certifications/{product_category}": {
      "get": {
        "summary": "Certifications",
        "operationId": "certifications_v1_certifications__product_category__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "product_category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Category"
            }
          },
          {
            "name": "domestic_supply",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domestic Supply"
            }
          },
          {
            "name": "exemption_claimed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exemption Claimed"
            }
          },
          {
            "name": "district_in_current_annexure",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District In Current Annexure"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/feedback": {
      "post": {
        "summary": "Feedback",
        "operationId": "feedback_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlliedResponse": {
        "properties": {
          "is_number": {
            "type": "string",
            "title": "Is Number"
          },
          "allied_standards": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/AlliedStandard"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Allied Standards"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "is_number",
          "allied_standards",
          "evidence"
        ],
        "title": "AlliedResponse"
      },
...
      "DirectoryResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StandardResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "DirectoryResponse"
      },
...
      "FeedbackResponse": {
        "properties": {
          "feedback_id": {
            "type": "string",
            "format": "uuid",
            "title": "Feedback Id"
          },
          "recommendation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recommendation Id"
          },
          "stored": {
            "type": "boolean",
            "title": "Stored"
          }
        },
        "type": "object",
        "required": [
          "feedback_id",
          "recommendation_id",
          "stored"
        ],
        "title": "FeedbackResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "HistoryResponse"
      },
...
      "RecommendResponse": {
        "properties": {
          "recommendation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recommendation Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "query_text": {
            "type": "string",
            "title": "Query Text"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "primary_standards": {
            "items": {
              "$ref": "#/components/schemas/PrimaryStandard"
            },
            "type": "array",
            "title": "Primary Standards"
          },
          "allied_standards": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/AlliedStandard"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Allied Standards"
          },
          "certification_requirements": {
            "items": {
              "$ref": "#/components/schemas/CertificationRequirement"
            },
            "type": "array",
            "title": "Certification Requirements"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Warning"
            },
            "type": "array",
            "title": "Warnings"
          },
          "normalization": {
            "additionalProperties": true,
            "type": "object",
            "title": "Normalization"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "recommendation_id",
          "timestamp",
          "query_text",
          "status",
          "message",
          "primary_standards",
          "allied_standards",
          "certification_requirements",
          "warnings",
          "normalization"
        ],
        "title": "RecommendResponse"
      },
      "StandardResponse": {
        "properties": {
          "record": {
            "additionalProperties": true,
            "type": "object",
            "title": "Record"
          },
          "version_status": {
            "$ref": "#/components/schemas/VersionInfo"
          },
          "certification_requirements": {
            "items": {
              "$ref": "#/components/schemas/CertificationRequirement"
            },
            "type": "array",
            "title": "Certification Requirements"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Warning"
            },
            "type": "array",
            "title": "Warnings"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "record",
          "version_status",
          "certification_requirements",
          "warnings",
          "evidence"
        ],
        "title": "StandardResponse"
      },
      "SystemResponse": {
        "properties": {
          "kb_fingerprint": {
            "type": "string",
            "title": "Kb Fingerprint"
          },
          "visible_records": {
            "type": "integer",
            "title": "Visible Records"
          },
          "verified_records": {
            "type": "integer",
            "title": "Verified Records"
          },
          "synthetic_records": {
            "type": "integer",
            "title": "Synthetic Records"
          },
          "include_synthetic": {
            "type": "boolean",
            "title": "Include Synthetic"
          },
          "confidence_threshold": {
            "type": "number",
            "title": "Confidence Threshold"
          },
          "graph_backend": {
            "type": "string",
            "title": "Graph Backend"
          },
          "models": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Models"
          },
          "ranking_adjustments": {
            "type": "string",
            "title": "Ranking Adjustments"
          },
          "notice": {
            "type": "string",
            "title": "Notice"
          }
        },
        "type": "object",
        "required": [
          "kb_fingerprint",
          "visible_records",
          "verified_records",
          "synthetic_records",
          "include_synthetic",
          "confidence_threshold",
          "graph_backend",
          "models",
          "ranking_adjustments",
          "notice"
        ],
        "title": "SystemResponse"
      },
...
      "RecommendRequest": {
        "additionalProperties": false,
        "properties": {
          "text": {
            "maxLength": 100000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "top_k": {
            "default": 5,
            "maximum": 10,
            "minimum": 1,
            "title": "Top K",
            "type": "integer"
          },
          "tender": {
            "default": false,
            "title": "Tender",
            "type": "boolean"
          },
          "language_hint": {
            "anyOf": [
              {
                "enum": [
                  "en",
                  "hi",
                  "hi-Latn",
                  "bn",
                  "mr",
                  "te",
                  "ta",
                  "gu",
                  "kn",
                  "ml",
                  "pa",
                  "ur"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Language Hint"
          },
          "product_category": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9_-]{1,80}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Product Category"
          },
          "certification_context": {
            "$ref": "#/components/schemas/CertificationContext"
          }
        },
        "required": [
          "text"
        ],
        "title": "RecommendRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}